demo: add a simple v307 blinky example and rename the original embassy_blinky
